==> compile_and_filter_dataset/1_getting_movie_discussions.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import praw

import re
import string

pd.options.display.max_colwidth = 75

### Using PRAW library as a wrapper for the Reddit API

with open('../.secret/ZSDSFI_client_id.txt') as f:
    client_id = f.read()

with open('../.secret/ZSDSFI_client_secret.txt') as f:
    client_secret = f.read()

reddit = praw.Reddit(
    client_id=client_id,
    client_secret=client_secret,
    user_agent="Movie Scraper by ZSDSFI"
)

I'm using the subreddit *r/discussionarchive* to get a list of **official** movie discussions on *r/movies*.

posts = []

movie_archive = reddit.subreddit("discussionarchive").new(limit=1000)

for submission in movie_archive:
    posts.append((submission.title, submission.url))

Reddit has a limitation of 1000 posts per subreddit. This shouldn't be an issue for this project. I have access to movie discussions from February 2016 to July 2022. And each discussion itself contains many comments. That's a lot of data.

df = pd.DataFrame(posts)
df.columns = ['post_title', 'url']

df

df['post_title'] = df['post_title'].apply(lambda x: x.lower())

df['id'] = df['url'].apply(praw.models.Submission.id_from_url)

# There are discussions for award shows and end-of-year discussions. We don't want those.

df[
    df['post_title'].str.contains('golden globes') |
    df['post_title'].str.contains('oscars') |
    df['post_title'].str.contains('best of') |
    ~df['post_title'].str.contains('discussion')
    ]

mask_out = df[
    df['post_title'].str.contains('golden globes') |
    df['post_title'].str.contains('oscars') |
    df['post_title'].str.contains('best of') |
    ~df['post_title'].str.contains('discussion')
    ].iloc[2:].index

df = df.drop(mask_out)

# Any duplicates?

df[df.duplicated(keep=False)]

df = df.drop_duplicates(keep='first')

### Using ReGex to get movie names from each post.
This will take a bit of grunt work because r/movies doesn't format each of its discussions the exact same way. Thankfully there is enough of a pattern that I can still do this in a few steps.

df['post_title']

title_Series = df['post_title'].rename('title')

title_Series =  title_Series.apply(lambda x: re.sub("(official discussion(: | - )| \[spoilers])", "", x))

title_Series[
    title_Series.str.contains('spoilers')|
    title_Series.str.contains('discussion')|
    title_Series.str.contains('international')
    ]

* International releases should be treated as normal.
* "Spoilers" is sometimes written in different formats.

title_Series = title_Series.apply(lambda x: x.replace(" (spoilers)", ""))
title_Series = title_Series.apply(lambda x: x.replace(" [spoilers", ""))
title_Series = title_Series.apply(lambda x: x.replace(" (international release)", ""))
title_Series = title_Series.apply(lambda x: x.replace(" (international thread)", ""))

title_Series[
    title_Series.str.contains('spoilers')|
    title_Series.str.contains('discussion')|
    title_Series.str.contains('international')
    ]

pattern = "((official international discussion)|(official international release discussion thread)|(official international release discussion))(: | - )"

title_Series = title_Series.apply(lambda x: re.sub(pattern=pattern, repl="", string=x))

title_Series[
    title_Series.str.contains('spoilers')|
    title_Series.str.contains('discussion')|
    title_Series.str.contains('international')
    ]

# Remove row 900

df['title'] = title_Series

df = df[['title', 'post_title', 'id', 'url']]

df = df.drop(index=900)

# A few more quick fixes

df[df['title'].str.contains('us thread')]

df[df['title'].str.contains('us release')]

df['title'] = df['title'].apply(lambda x: x.replace(" (us thread)", ""))
df['title'] = df['title'].apply(lambda x: x.replace(" (us release)", ""))

df[df['post_title'].str.contains('discusion')]

df['title'].loc[42] = 'apollo 10½: a space age childhood'

df[df['post_title'].str.contains('american assassin')]

df['title'].loc[795] = 'american assassin'

df.sample(5)

**Note that some movies have a year in the title because they are remakes. I'm removing the year in order to make matching easier if I try to combine this dataset with a ratings dataset (like from IMDb).**

df[df['title'].str.contains(pat="\(\d{4}\)", regex=True)]

# Remove year

df['title'] = df['title'].apply(
    lambda x: re.sub(pattern=r" \(\d{4}\)", repl="", string=x)
)

**Note that we now have some duplicates in the "title" column since some movies have separate US discussions and an international discussions. I will leave this as is for this notebook, since they are still separate Reddit threads. I will do more work on this issue in a future notebook, when I clean up commments.**

df[df.duplicated(subset=['title'], keep=False)]

# Run this once
# df.to_csv("../data/movies_cleaned.csv", index=False)

pd.read_csv("../data/movies_cleaned.csv")
