==> compile_and_filter_dataset/2_getting_comments_json.ipynb <==
import numpy as np
import pandas as pd
from collections import OrderedDict

import praw

import json

from datetime import datetime

pd.options.display.max_colwidth = 75

df = pd.read_csv('../data/movies_cleaned.csv', index_col='id')

with open('../.secret/ZSDSFI_client_id.txt') as f:
    client_id = f.read()

with open('../.secret/ZSDSFI_client_secret.txt') as f:
    client_secret = f.read()

reddit = praw.Reddit(
    client_id=client_id,
    client_secret=client_secret,
    user_agent="Movie Scraper by ZSDSFI"
)

df.head()

all_ids = df.index.tolist()

# The following cell can about one hour to run.

movies_comments = {}

for id in all_ids:
    submission = reddit.submission(id=id)
    comments = []
    # Using a try-except block because at least one movie causes an error, as the thread no longer exists 
    try:
        submission.comment_sort = 'top'
        post_date_utc = submission.created_utc
        for top_level_comment in submission.comments[:100]:
            comments.append(top_level_comment.body)
    except:
        continue
    movies_comments[id] = {
        'title': df.loc[id]['title'],
        'comments': comments,
        'post_date_utc': post_date_utc
    }

pd.DataFrame(movies_comments).T.info()

pd.DataFrame(movies_comments).T.sample(10)

# Saving the lists of movie comments as a JSON file
# Run this once

# with open("../data/movies_comments.json", "w") as outfile:
#     json.dump(movies_comments, outfile)
