==> compile_and_filter_dataset/3_cleaning_comments.ipynb <==
import numpy as np
import pandas as pd

import re
import string
from datetime import datetime

import praw

import json

pd.options.display.max_colwidth = 100
pd.options.display.max_rows = 100
seed = 55
np.random.seed(seed)

# Reading in list of movies

movies_cleaned_df = pd.read_csv('../data/movies_cleaned.csv', index_col='id')

# Reading in the comments I collected

with open('../data/movies_comments.json') as f:
    movies_comments_dict = json.load(f)

# Making the comments collection a DataFrame

movies_comments_df = pd.DataFrame(movies_comments_dict).T

movies_comments_df

# Combining international and domestic releases (same title; different discussion thread)

movies_comments_df[movies_comments_df['title'].duplicated(keep=False)]

# The Girl on the Train is not a duplicate. Two movies with the same name.

multi_discussion_movies = \
    movies_comments_df[
        movies_comments_df['title'].duplicated()
        ]['title'].drop(index='56973x').tolist()

multi_discussion_index = \
    movies_comments_df[
        movies_comments_df['title'].duplicated(keep=False)
        ].drop(index=['56973x', 'ltdr45']).index.tolist()

movies_comments_df[movies_comments_df['title'].isin(multi_discussion_movies)]['title'].value_counts()

for title in multi_discussion_movies:
    duo = movies_comments_df[movies_comments_df['title'] == title]
    discussion_1 = duo.index[0]
    discussion_2 = duo.index[1]
    movies_comments_df.loc[discussion_1]['comments'] +=\
        movies_comments_df.loc[discussion_2]['comments']
    movies_comments_df = movies_comments_df.drop(index=discussion_2)

# Double checking if there are more duplicates

movies_comments_df\
    [(movies_comments_df['title']\
        .apply(lambda x: x.replace("the ", "") if x.startswith("the ") else x))\
            .apply(lambda x: x[:8]).duplicated(keep=False)]

# Still need to merge these movies:

# - star wars: the last jedi -- 7jwxnd, 7rb3uy
# - avengers: infinity war -- 8f84h0, 8gvr6n
# - jeen-yuhs -- sv7agc, t7pqlt
# - avengers: endgame -- bh8iei, bk33kl
# - they shall not grow old -- 9x244z, a5yk8k
# - thor: ragnarok -- 7agfes, 78ivjl

# star wars: the last jedi -- 7jwxnd, 7rb3uy

movies_comments_df.loc[['7jwxnd', '7rb3uy']]

# Merging the threads

movies_comments_df.loc['7jwxnd']['comments']\
     += movies_comments_df.loc['7rb3uy']['comments']

movies_comments_df = movies_comments_df.drop(index=['7rb3uy'])

# avengers: infinity war -- 8f84h0, 8gvr6n

movies_comments_df.loc[['8f84h0', '8gvr6n']]

# Removing administrative comments

movies_comments_df.loc['8f84h0']['comments']\
    = movies_comments_df.loc['8f84h0']['comments'][1:]

movies_comments_df.loc['8gvr6n']['comments']\
    = movies_comments_df.loc['8gvr6n']['comments'][1:]

# Merging the threads

movies_comments_df.loc['8f84h0']['comments']\
     += movies_comments_df.loc['8gvr6n']['comments']
     
movies_comments_df = movies_comments_df.drop(index=['8gvr6n'])

# jeen-yuhs -- sv7agc, t7pqlt

# This is more of a three-part docuseries than a movie. 
# Plus, Part 2 is missing from the dataset. 
# Better to drop it. Sorry, Kanye fans.

movies_comments_df = movies_comments_df.drop(['sv7agc', 't7pqlt'])

# avengers: endgame -- bh8iei, bk33kl

movies_comments_df.loc[['bh8iei', 'bk33kl']]

# Removing administrative comments

movies_comments_df.loc['bh8iei']['comments']\
    = movies_comments_df.loc['bh8iei']['comments'][1:]

# Merging the threads

movies_comments_df.loc['bh8iei']['comments'] +=\
    movies_comments_df.loc['bk33kl']['comments']

movies_comments_df = movies_comments_df.drop(index=['bk33kl'])

# they shall not grow old -- 9x244z, a5yk8k

movies_comments_df.loc[['9x244z', 'a5yk8k']]

# Removing administrative comments
movies_comments_df.loc['a5yk8k']['comments']\
    = movies_comments_df.loc['a5yk8k']['comments'][1:]

# Merging the threads

movies_comments_df.loc['a5yk8k']['comments'] +=\
    movies_comments_df.loc['9x244z']['comments']

movies_comments_df = movies_comments_df.drop(index=['9x244z'])

# - thor: ragnarok -- 7agfes, 78ivjl

movies_comments_df.loc[['7agfes', '78ivjl']]

# Removing administrative comments
movies_comments_df.loc['7agfes']['comments']\
    = movies_comments_df.loc['7agfes']['comments'][1:]

# Merging the threads

movies_comments_df.loc['7agfes']['comments'] +=\
    movies_comments_df.loc['78ivjl']['comments']

movies_comments_df = movies_comments_df.drop(index=['78ivjl'])

# correct spelling

movies_comments_df.loc['7agfes']['title'] = 'thor: ragnarok'

# correcting one more movie title:
movies_comments_df.loc['ncomky']

movies_comments_df.loc['ncomky']['title'] = 'army of the dead'

# Supposedly all movies have been merged now

# Adding date columns

movies_comments_df['post_year'] = movies_comments_df['post_date_utc'].apply(lambda x: datetime.fromtimestamp(x).year)
movies_comments_df['post_month'] = movies_comments_df['post_date_utc'].apply(lambda x: datetime.fromtimestamp(x).month)
movies_comments_df['post_day'] = movies_comments_df['post_date_utc'].apply(lambda x: datetime.fromtimestamp(x).day)

movies_comments_df.sample(3)

# Making a CSV of JUST the movie title, id, and date it was discussed on Reddit.
# This is will be helpful when trying to match movies to their counterparts in a ratings dataset (like IMDb).

reddit_movies_final = movies_comments_df.drop(columns=['comments', 'post_date_utc']).reset_index()
reddit_movies_final.rename(columns={'index': 'id'}, inplace=True)

reddit_movies_final.sample(10)

# Run this once
# reddit_movies_final.to_csv("../data/reddit_movies_final.csv", index=True)

######################

# Exploding the comments

comments_df = movies_comments_df.explode('comments').reset_index()
comments_df.rename(columns={'index': 'id'}, inplace=True)
comments_df.sample(5)

# Check comments for duplicates
# Manually review every comment where the first 50 characters are appear elsewhere.
# Could be: spam, deleted, removed
# Also could be same commenter in two threads about the same movie

comment_counts = comments_df['comments'].apply(lambda x: x[:50]).value_counts()
maybe_duplicate_comments = comment_counts[comment_counts > 1].keys()
maybe_duplicate_comments = list(maybe_duplicate_comments)
len(maybe_duplicate_comments)

# Review the comments in maybe_duplicate_comments. 
# Decide which are actually duplicates, which could be coincidental, 
# and which should only be kept in one thread (same movie)

list(enumerate(maybe_duplicate_comments))

# Manual review, since there are only 65

checking = -1

# Run this cell multiple times to check for duplicates

checking += 1
print(checking)

comments_df[
    comments_df['comments'].apply(lambda x: x[:50])\
         == maybe_duplicate_comments[checking]
    ]

drop_all = [
    0, 1, 4, 12, 21, 41, 42, 43, 46, 52, 
    ]

keep_one = [
    9, 14, 18, 19, 22, 25, 26, 31, 32, 33, 34, 37, 39, 40, 
    44, 47, 49, 56, 57, 
    ]

do_nothing = [
    2, 3, 5, 6, 7, 8, 10, 11, 13, 15, 16, 17, 20, 23, 24, 
    27, 28, 29, 30, 35, 36, 38, 45, 48, 50, 51, 53, 54, 55,
    58, 59, 60, 61, 62, 63, 64
    ]

i_to_drop = []

for i, comment in enumerate(maybe_duplicate_comments):
    if i in drop_all:
        i_to_drop.extend(comments_df[comments_df['comments'].apply(lambda x: x[:50]) == comment].index.tolist())
    elif i in keep_one: 
        i_to_drop.extend(comments_df[comments_df['comments'].apply(lambda x: x[:50]) == comment].index.tolist()[1:])

comments_df = comments_df.drop(index=i_to_drop)

##################
# Checking for other administrative comments
# They are usually the first comment

df_only_first_comment = comments_df.drop_duplicates(subset='id', keep='first')

# Checking for common phrases that would be in an administrative comment.

mask = (
    df_only_first_comment['comments'].apply(lambda x: 'heads up' in x.lower())\
        | df_only_first_comment['comments'].apply(lambda x: 'FYI' in x.lower())
            | df_only_first_comment['comments'].apply(lambda x: 'r/movies' in x.lower())\
                | df_only_first_comment['comments'].apply(lambda x: 'pinned' in x.lower())
    |                df_only_first_comment['comments'].apply(lambda x: 'AMA' in x)
)

df_only_first_comment[mask]

# Only the first one is an actual comment about a movie. That was some good search criteria.

df_only_first_comment[mask].index[1:]

comments_df = comments_df.drop(
    df_only_first_comment[mask].index[1:]
    )

# Run this once
# comments_df.to_csv("../data/comments_exploded.csv", index=True)

pd.read_csv("../data/comments_exploded.csv", index_col=0)
