==> compile_and_filter_dataset/5_get_sentiments.ipynb <==
# Porportional sentiment of each Reddit r/movies thread

This was a last minute add-on in order to feature-engineer sentiment labels the Reddit comments I'm using. 

This was run in Google Colab to make use of the Cloud GPU capability, as running it on my machine was far too slow.

Following instructions from [this notebook from Google](https://colab.research.google.com/notebooks/pro.ipynb).

gpu_info = !nvidia-smi
gpu_info = '\n'.join(gpu_info)
if gpu_info.find('failed') >= 0:
  print('Not connected to a GPU')
else:
  print(gpu_info)

from psutil import virtual_memory
ram_gb = virtual_memory().total / 1e9
print('Your runtime has {:.1f} gigabytes of available RAM\n'.format(ram_gb))

if ram_gb < 20:
  print('Not using a high-RAM runtime')
else:
  print('You are using a high-RAM runtime!')

!pip install transformers

import time
start_time = time.time()
def time_check(start=None):
    if start:
        t = time.time() - start
    else:
        t = time.time() - start_time
    print(f'Time check: {t//60:.0f} minutes and {t%60:.0f} seconds')

import pandas as pd
import numpy as np
import tensorflow as tf
from transformers import pipeline
from google.colab import files

tf.test.gpu_device_name()

# Read in dataset

data_url = "https://raw.githubusercontent.com/zshoorbajee/reddit-movie-comments-nlp/main/data/data_final.csv"
df = pd.read_csv(data_url)

df.head()

### "Imploding" the dataframe

Currently, each comment is its own row, giving the dataset over 70,000 rows. I need to ge the proportion of positive and negative comments from each movie. So it would be easier to treat each movie as one row.

# "Implode" comments so each movie is one row

comments_imploded = df.groupby('id')['comments'].agg(list)
df = df.drop_duplicates(subset='id').drop(columns=['comments'])
df = df.join(comments_imploded, on='id')
df['n_comments'] = df['comments'].apply(len)
df.head()

### Get sentiment label of each comment.
I'm using the Hugging Face Transformers library with a commonly used sentiment analysis model trained on 58 million tweets. The model is called [cardiffnlp/twitter-roberta-base-sentiment](https://huggingface.co/cardiffnlp/twitter-roberta-base-sentiment).

# Commonly used sentiment analysis model from HuggingFace Transformers  

sentiment_pipeline = pipeline(model="cardiffnlp/twitter-roberta-base-sentiment")

# Function to get a proportion of sentiment labels for each comment thread.

def get_sentiments(comments_list, raw_counts=False):
  """
  Takes in a list of comments as strings.
  Returns the a dictionary with the proportion of positive, 
  negative, and neutral comments based on a Transformers model.
  """
  sentiments_count = {"negative": 0, "neutral": 0, "positive": 0}
  comment_count = 0

  for comment in comments_list:
    try:
      sentiment = sentiment_pipeline(comment)
    except:
      continue
    label = sentiment[0]['label']
    if label == "LABEL_0":
      sentiments_count["negative"] += 1
    elif label == "LABEL_1":
      sentiments_count["neutral"] += 1
    elif label == "LABEL_2":
      sentiments_count["positive"] += 1
    comment_count += 1

  if raw_counts:
    return sentiments_count
  else:
    sentiments_normalized = {
      "negative": sentiments_count["negative"] / comment_count,
      "neutral": sentiments_count["neutral"] / comment_count,
      "positive": sentiments_count["positive"] / comment_count
      }
    return sentiments_normalized

# Test the function:

example_sentiments = df['comments'].head().apply(get_sentiments)
example_sentiments

# Confirmed that it works.

df['sentiments_norm'] = df['comments'].apply(get_sentiments)

time_check()

df['neg_norm'] = df['sentiments_norm'].apply(lambda x: x['negative'])
df['ntrl_norm'] = df['sentiments_norm'].apply(lambda x: x['neutral'])
df['pos_norm'] = df['sentiments_norm'].apply(lambda x: x['positive'])

reddit_movie_sentiments = df[['id', 'title', 'neg_norm', 'ntrl_norm', 'pos_norm']]

reddit_movie_sentiments.to_csv("reddit_movie_sentiments.csv", header=True, index=False)

files.download('reddit_movie_sentiments.csv')
